// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  patient: 'patient',
  provider: 'provider',
  pharmacy: 'pharmacy',
  admin: 'admin'
} as constpassword

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const VerificationStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected'
} as const

export type VerificationStatus = (typeof VerificationStatus)[keyof typeof VerificationStatus]


export const ConsultationStatus = {
  pending: 'pending',
  scheduled: 'scheduled',
  completed: 'completed',
  cancelled: 'cancelled',
  rejected: 'rejected',
  expired: 'expired'
} as const

export type ConsultationStatus = (typeof ConsultationStatus)[keyof typeof ConsultationStatus]